basic user-service 60% complete (detail exist)
The attendance service is now complete, with Flask and Spring Boot effectively communicating with each other. In this setup, Flask manages the attendance function while Spring Boot handles everything else. Attendance is based on GPS coordinates and lecture schedules. If the GPS coordinates or lecture schedules are incorrect, students will not be able to attend the lecture.